Rename nvdecode module to dataloader
Explicitly tell what contains this module from a user perspective
